Smoothen out the reported current
By smoothening the reported current we can avoid the disengage when the wheel is engaged. When we're way off the line it often starts with maxPWM which draws a ton of current. This code approach allows that initial spike to be flattened. This way the disengage % could be set to much lower.